# app/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ParaGraph Backend",
    "description": "ParaGraph workflow backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/executions/compile": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "Compile Workflow",
        "operationId": "compile_workflow_executions_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileWorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileWorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "Start Execution",
        "operationId": "start_execution_executions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartExecutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions/{run_id}": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "Get Execution",
        "operationId": "get_execution_executions__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionRunState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions/{run_id}/events": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "Get Execution Events",
        "operationId": "get_execution_events_executions__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions/{run_id}/cancel": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "Cancel Execution",
        "operationId": "cancel_execution_executions__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/executions/{run_id}/resume": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "Resume Execution",
        "operationId": "resume_execution_executions__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompileWorkflowRequest": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/WorkflowDefinition"
          }
        },
        "type": "object",
        "required": [
          "definition"
        ],
        "title": "CompileWorkflowRequest"
      },
      "CompileWorkflowResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "diagnostics": {
            "items": {
              "$ref": "#/components/schemas/CompilerDiagnostic"
            },
            "type": "array",
            "title": "Diagnostics"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompiledExecutionPlan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "CompileWorkflowResponse"
      },
      "CompiledExecutionPlan": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 2
          },
          "step_order": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Step Order"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ExecutionStepPlan"
            },
            "type": "array",
            "title": "Steps"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "plan_id"
        ],
        "title": "CompiledExecutionPlan"
      },
      "CompilerDiagnostic": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "level": {
            "type": "string",
            "enum": [
              "error",
              "warning"
            ],
            "title": "Level",
            "default": "error"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "connection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowConnection"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "CompilerDiagnostic"
      },
...
      "EventHistoryResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ExecutionEventEnvelope"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "EventHistoryResponse"
      },
      "ExecutionActionResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled",
              "paused"
            ],
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "message"
        ],
        "title": "ExecutionActionResponse"
      },
      "ExecutionBinding": {
        "properties": {
          "binding_type": {
            "type": "string",
            "enum": [
              "input",
              "controller"
            ],
            "title": "Binding Type",
            "default": "input"
          },
          "input_name": {
            "type": "string",
            "title": "Input Name"
          },
          "source_node_id": {
            "type": "string",
            "title": "Source Node Id"
          },
          "source_output": {
            "type": "string",
            "title": "Source Output"
          }
        },
        "type": "object",
        "required": [
          "input_name",
          "source_node_id",
          "source_output"
        ],
        "title": "ExecutionBinding"
      },
      "ExecutionEventEnvelope": {
        "properties": {
          "event_type": {
            "type": "string",
            "enum": [
              "execution.queued",
              "execution.started",
              "execution.step.started",
              "execution.step.progress",
              "execution.step.completed",
              "execution.step.failed",
              "execution.cancellation.requested",
              "execution.cancelled",
              "execution.step.retry.started",
              "execution.step.retry.failed",
              "execution.step.timeout",
              "execution.paused",
              "execution.resumed",
              "execution.recovered",
              "execution.completed",
              "execution.failed"
            ],
            "title": "Event Type"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "step_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_type",
          "run_id",
          "sequence"
        ],
        "title": "ExecutionEventEnvelope"
      },
      "ExecutionRunState": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "execution_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Session Id"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled",
              "paused"
            ],
            "title": "Status",
            "default": "queued"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ExecutionStepState"
            },
            "type": "array",
            "title": "Steps"
          },
          "outputs": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Outputs"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "pause_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pause Payload"
          },
          "resume_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Token"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompiledExecutionPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "cancellation_requested": {
            "type": "boolean",
            "title": "Cancellation Requested",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "plan_id"
        ],
        "title": "ExecutionRunState"
      },
      "ExecutionStepPlan": {
        "properties": {
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "node_version": {
            "type": "integer",
            "title": "Node Version"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "executor_key": {
            "type": "string",
            "title": "Executor Key"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "bindings": {
            "items": {
              "$ref": "#/components/schemas/ExecutionBinding"
            },
            "type": "array",
            "title": "Bindings"
          },
          "timeout_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Ms"
          },
          "retries": {
            "type": "integer",
            "title": "Retries",
            "default": 0
          },
          "cacheable": {
            "type": "boolean",
            "title": "Cacheable",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "step_id",
          "node_id",
          "node_type",
          "node_version",
          "category",
          "executor_key"
        ],
        "title": "ExecutionStepPlan"
      },
      "ExecutionStepState": {
        "properties": {
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "skipped"
            ],
            "title": "Status",
            "default": "queued"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "output": {
            "additionalProperties": true,
            "type": "object",
            "title": "Output"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "pause_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pause Payload"
          },
          "resume_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Token"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "default": 0
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count",
            "default": 0
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          }
        },
        "type": "object",
        "required": [
          "step_id",
          "node_id",
          "node_type"
        ],
        "title": "ExecutionStepState"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeExecutionRequest": {
        "properties": {
          "resume_token": {
            "type": "string",
            "title": "Resume Token"
          },
          "reviewed_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Payload"
          }
        },
        "type": "object",
        "required": [
          "resume_token"
        ],
        "title": "ResumeExecutionRequest"
      },
      "StartExecutionRequest": {
        "properties": {
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          },
          "execution_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Session Id"
          },
          "plan": {
            "$ref": "#/components/schemas/CompiledExecutionPlan"
          }
        },
        "type": "object",
        "required": [
          "plan"
        ],
        "title": "StartExecutionRequest"
      },
      "StartExecutionResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled",
              "paused"
            ],
            "title": "Status"
          },
          "execution_session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Session Id"
          },
          "poll_interval": {
            "type": "number",
            "title": "Poll Interval",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "StartExecutionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowConnection": {
        "properties": {
          "from_node": {
            "type": "string",
            "title": "From Node"
          },
          "to_node": {
            "type": "string",
            "title": "To Node"
          },
          "connection_type": {
            "type": "string",
            "enum": [
              "data",
              "controller"
            ],
            "title": "Connection Type",
            "default": "data"
          },
          "from_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Output"
          },
          "to_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Input"
          },
          "from_controller": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Controller"
          },
          "to_controller": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Controller"
          }
        },
        "type": "object",
        "required": [
          "from_node",
          "to_node"
        ],
        "title": "WorkflowConnection"
      },
      "WorkflowDefinition": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 2,
            "title": "Schema Version"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNodeInstance"
            },
            "type": "array",
            "title": "Nodes"
          },
          "connections": {
            "items": {
              "$ref": "#/components/schemas/WorkflowConnection"
            },
            "type": "array",
            "title": "Connections"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "schema_version"
        ],
        "title": "WorkflowDefinition"
      },
...
      "WorkflowNodeInstance": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "node_type": {
            "type": "string",
            "title": "Node Type"
          },
          "node_version": {
            "type": "integer",
            "title": "Node Version"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "timeout_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Ms"
          },
          "retries": {
            "type": "integer",
            "title": "Retries",
            "default": 0
          },
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "node_type",
          "node_version"
        ],
        "title": "WorkflowNodeInstance"
      },
...
    }
  }
}